[Hunter] Highly experimental Withering Fire
Since the PTR doesn't look to be coming up this is made with a very large amount of assumptions.

This implementation assumes that if withering fire is active, then a deathblow proc will not reset the cooldown of the Black Arrow action. Instead it will do the following:
Fire a regular Black Arrow towards the main target, this black arrow can do everything that a regular black arrow can do.
Fire up to 2 secondary black arrows at targets that are NOT the main target, these arrows apply the black arrow dot and that functions exactly like a regular black arrow dot.